Add "Unit" unit (#34)
* Add new unit "unit"

This commit looks to add a new unit to be supported. HelloFresh in
particular often uses "unit" when referencing an atomic item. For
example "1 unit apple" to mean "1 apple".

* Add unit to MISINTERPRETED_UNITS

This was being interpreted by pint as a "micronit" or a "micro newton
inch". But instead we want it to represent itself -- "unit".

---------

Co-authored-by: tom <tpstrange@gmail.com>