Refactor: Refactor webtools labeller search functionality.
Moves the different search types into separate functions (in search.py).
For string search, create a user defined SQL function to finds sublist matches of a target list of tokens in the tokens lists stored in the database.
The target list of tokens is obtained by running the user input string through the PreProcessor function.
This lets us avoid a lot of regex shenanigans we had before, and also fixes some bugs where search for strings like "medium-sized" or "skin-on" would return no results.

This also simplifies the labeller_search function in app.py, which should make it more maintainable.